+++ pages/mypage/index.tsx
@@ -2,7 +2,7 @@ import Layout from '@components/Layout'
 import React from 'react'
 
 export default function MyPage() {
-  return <div>에고</div>
+  return <div></div>
 }
 
 MyPage.getLayout = function getLayout(page: React.ReactElement) {
